refactor: Extract code remediation methods into CodeRemediationService
- Created src/tools/modules/services/code-remediation-service.ts (369 lines)
- Extracted all remediation and refactoring helper methods:
  - applySecurityFixes, applyPerformanceFixes, applyMemoryLeakFixes
  - detectAllIssues, applyBestPracticesFixes, applyTypeSafetyFixes
  - refactorForPerformance, refactorForMaintainability, refactorForAccessibility
  - refactorForTypeSafety, refactorToModernPatterns, refactorComprehensive
  - generateRefactoredTests
- All methods now static in CodeRemediationService class
- Updated all method calls in tools/index.ts to use CodeRemediationService
- Removed old private method implementations (360 lines)

Impact:
- tools/index.ts reduced from 2,606 to 2,246 lines (360 lines removed)
- All 296 tests passing ✅
- No TypeScript errors ✅

Progress: File reduced from original 5,169 to 2,246 lines (56.5% reduction)
Remaining: Need to remove ~1,246 more lines to reach < 1,000 target
